+++ src/main/java/com/avp/common/block/RadiatedBlock.java
@@ -3,7 +3,6 @@ package com.avp.common.block;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
@@ -15,7 +14,6 @@ import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.NotNull;
 
 import com.avp.common.effect.AVPEffects;
-import com.avp.common.item.AVPItemTags;
 import com.avp.common.util.AVPPredicates;
 
 public class RadiatedBlock extends Block {
@@ -36,15 +34,11 @@ public class RadiatedBlock extends Block {
 
     @Override
     public void stepOn(Level level, BlockPos blockPos, BlockState blockState, Entity entity) {
-        if (entity instanceof LivingEntity livingEntity && !AVPPredicates.IS_IMMORTAL.test(livingEntity)) {
-            var armorCheck = livingEntity.getItemBySlot(EquipmentSlot.HEAD).is(AVPItemTags.RADIATION_RESISTANT_ARMOR) &&
-                livingEntity.getItemBySlot(EquipmentSlot.CHEST).is(AVPItemTags.RADIATION_RESISTANT_ARMOR) &&
-                livingEntity.getItemBySlot(EquipmentSlot.LEGS).is(AVPItemTags.RADIATION_RESISTANT_ARMOR) &&
-                livingEntity.getItemBySlot(EquipmentSlot.FEET).is(AVPItemTags.RADIATION_RESISTANT_ARMOR);
-            if (!armorCheck && !livingEntity.hasEffect(AVPEffects.RADIATION_EFFECT)) {
-                livingEntity.addEffect(new MobEffectInstance(AVPEffects.RADIATION_EFFECT, Integer.MAX_VALUE, 0));
-            }
+        if (AVPPredicates.canBeIrradiated(entity) && entity instanceof LivingEntity livingEntity) {
+            // Apply radiation effect.
+            livingEntity.addEffect(new MobEffectInstance(AVPEffects.RADIATION_EFFECT, Integer.MAX_VALUE, 0));
         }
+
         super.stepOn(level, blockPos, blockState, entity);
     }
 }