+++ src/main/java/com/avp/common/block/resin/IrradiatedResinBlock.java
@@ -3,13 +3,11 @@ package com.avp.common.block.resin;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 
 import com.avp.common.effect.AVPEffects;
-import com.avp.common.item.AVPItemTags;
 import com.avp.common.util.AVPPredicates;
 
 public class IrradiatedResinBlock extends ResinBlock {
@@ -20,15 +18,11 @@ public class IrradiatedResinBlock extends ResinBlock {
 
     @Override
     public void stepOn(Level level, BlockPos blockPos, BlockState blockState, Entity entity) {
-        if (entity instanceof LivingEntity livingEntity && !AVPPredicates.IS_IMMORTAL.test(livingEntity)) {
-            var armorCheck = livingEntity.getItemBySlot(EquipmentSlot.HEAD).is(AVPItemTags.RADIATION_RESISTANT_ARMOR) &&
-                livingEntity.getItemBySlot(EquipmentSlot.CHEST).is(AVPItemTags.RADIATION_RESISTANT_ARMOR) &&
-                livingEntity.getItemBySlot(EquipmentSlot.LEGS).is(AVPItemTags.RADIATION_RESISTANT_ARMOR) &&
-                livingEntity.getItemBySlot(EquipmentSlot.FEET).is(AVPItemTags.RADIATION_RESISTANT_ARMOR);
-            if (!armorCheck && !livingEntity.hasEffect(AVPEffects.RADIATION_EFFECT)) {
-                livingEntity.addEffect(new MobEffectInstance(AVPEffects.RADIATION_EFFECT, Integer.MAX_VALUE, 0));
-            }
+        if (AVPPredicates.canBeIrradiated(entity) && entity instanceof LivingEntity livingEntity) {
+            // Apply radiation effect.
+            livingEntity.addEffect(new MobEffectInstance(AVPEffects.RADIATION_EFFECT, Integer.MAX_VALUE, 0));
         }
+
         super.stepOn(level, blockPos, blockState, entity);
     }
 }