+++ src/main/java/com/avp/mixin/MixinChestBlock_RadioactiveChest.java
@@ -2,7 +2,6 @@ package com.avp.mixin;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.effect.MobEffectInstance;
-import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.ChestBlock;
@@ -19,8 +18,6 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import com.avp.common.effect.AVPEffects;
-import com.avp.common.entity.AVPEntityTypeTags;
-import com.avp.common.item.AVPItemTags;
 import com.avp.common.util.AVPPredicates;
 
 @Mixin(ChestBlock.class)
@@ -41,41 +38,20 @@ public abstract class MixinChestBlock_RadioactiveChest {
     @Unique
     private <T extends BlockEntity> BlockEntityTicker<T> createServerTicker(BlockEntityType<T> blockEntityType) {
         return (level, pos, state, blockEntity) -> {
-            if (blockEntity instanceof ChestBlockEntity chestBlockEntity && containsRadiationItems(chestBlockEntity)) {
+            if (blockEntity instanceof ChestBlockEntity chestBlockEntity && AVPPredicates.containsIrradiatedItems(chestBlockEntity)) {
                 applyRadiationEffect(level, pos);
             }
         };
     }
 
     @Unique
-    private static boolean containsRadiationItems(ChestBlockEntity chest) {
-        for (var i = 0; i < chest.getContainerSize(); i++) {
-            var stack = chest.getItem(i);
-            if (!stack.isEmpty() && stack.is(AVPItemTags.RADIATION_ITEMS)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    @Unique
     private static void applyRadiationEffect(Level level, BlockPos pos) {
         var effectRadius = new AABB(pos).inflate(3);
 
-        var nearbyEntities = level.getEntitiesOfClass(
-            LivingEntity.class,
-            effectRadius,
-            entity -> entity.isAlive() && !entity.getType().is(AVPEntityTypeTags.RADIATION_RESISTANT)
-        );
-
-        for (var livingEntity : nearbyEntities) {
-            var armorCheck = livingEntity.getItemBySlot(EquipmentSlot.HEAD).is(AVPItemTags.RADIATION_RESISTANT_ARMOR) &&
-                livingEntity.getItemBySlot(EquipmentSlot.CHEST).is(AVPItemTags.RADIATION_RESISTANT_ARMOR) &&
-                livingEntity.getItemBySlot(EquipmentSlot.LEGS).is(AVPItemTags.RADIATION_RESISTANT_ARMOR) &&
-                livingEntity.getItemBySlot(EquipmentSlot.FEET).is(AVPItemTags.RADIATION_RESISTANT_ARMOR);
-            if (!armorCheck && !AVPPredicates.IS_IMMORTAL.test(livingEntity)) {
-                livingEntity.addEffect(new MobEffectInstance(AVPEffects.RADIATION_EFFECT, Integer.MAX_VALUE, 0));
-            }
-        }
+        level.getEntitiesOfClass(LivingEntity.class, effectRadius, AVPPredicates::canBeIrradiated)
+            .forEach(target -> {
+                var mobEffectInstance = new MobEffectInstance(AVPEffects.RADIATION_EFFECT, Integer.MAX_VALUE, 0);
+                target.addEffect(mobEffectInstance);
+            });
     }
 }