+++ src/main/java/com/avp/mixin/MixinItem_GiveRads.java
@@ -1,7 +1,7 @@
 package com.avp.mixin;
 
+import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
@@ -11,7 +11,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import com.avp.common.entity.AVPEntityTypeTags;
+import com.avp.common.effect.AVPEffects;
 import com.avp.common.item.AVPItemTags;
 import com.avp.common.util.AVPPredicates;
 
@@ -21,22 +21,20 @@ public class MixinItem_GiveRads {
     @Inject(method = "inventoryTick", at = @At("HEAD"))
     private void giveRads(ItemStack stack, Level level, Entity entity, int slotId, boolean isSelected, CallbackInfo ci) {
         if (
-            !level.isClientSide() && stack.is(AVPItemTags.RADIATION_ITEMS) && entity instanceof LivingEntity livingEntity
-                && livingEntity.isAlive() && !livingEntity.getType().is(AVPEntityTypeTags.RADIATION_RESISTANT)
+            // Only want to run this logic server-side.
+            level.isClientSide()
+                // Only run this logic for radiation-emitting items.
+                || !stack.is(AVPItemTags.RADIATION_ITEMS)
+                // Only run this logic if the entity can be irradiated.
+                || !AVPPredicates.canBeIrradiated(entity)
+                // Sanity check + allow compiler to assert entity type to get livingEntity ref access.
+                || !(entity instanceof LivingEntity livingEntity)
         ) {
-            var armorCheck = livingEntity.getItemBySlot(EquipmentSlot.HEAD).is(AVPItemTags.RADIATION_RESISTANT_ARMOR) &&
-                livingEntity.getItemBySlot(EquipmentSlot.CHEST).is(AVPItemTags.RADIATION_RESISTANT_ARMOR) &&
-                livingEntity.getItemBySlot(EquipmentSlot.LEGS).is(AVPItemTags.RADIATION_RESISTANT_ARMOR) &&
-                livingEntity.getItemBySlot(EquipmentSlot.FEET).is(AVPItemTags.RADIATION_RESISTANT_ARMOR);
-            if (!armorCheck && !AVPPredicates.IS_IMMORTAL.test(livingEntity)) {
-                livingEntity.addEffect(
-                    new net.minecraft.world.effect.MobEffectInstance(
-                        com.avp.common.effect.AVPEffects.RADIATION_EFFECT,
-                        Integer.MAX_VALUE,
-                        0
-                    )
-                );
-            }
+            return;
         }
+
+        // Apply the radiation effect.
+        var mobEffectInstance = new MobEffectInstance(AVPEffects.RADIATION_EFFECT, Integer.MAX_VALUE, 0);
+        livingEntity.addEffect(mobEffectInstance);
     }
 }